Provide thrift 0.9.2 image
Hacky solution to provide a 0.9.2 image. This involves checking
the version number and choosing the domain based on that (using the
apache archive if the version is older that 0.9.3). This is a bad
solution, but I don't have time to do this properly. The correct way is
to make an onbuild base command which sets the repo appropriately